requirements.txt
aws-cdk-lib==2.89.0
constructs>=10.0.0,<11.0.0
boto3==1.34.136
requests>=2.32.3
cdk-nag